+++ health/health.h
@@ -35,16 +35,9 @@ extern unsigned int default_health_enabled;
 #define HEALTH_LISTEN_BACKLOG 4096
 #endif
 
-#define HEALTH_ALARM_KEY "alarm"
-#define HEALTH_TEMPLATE_KEY "template"
+
 #define HEALTH_ON_KEY "on"
-#define HEALTH_CONTEXT_KEY "context"
-#define HEALTH_CHART_KEY "chart"
-#define HEALTH_HOST_KEY "hosts"
-#define HEALTH_OS_KEY "os"
-#define HEALTH_FAMILIES_KEY "families"
-#define HEALTH_LOOKUP_KEY "lookup"
-#define HEALTH_CALC_KEY "calc"
+
 #define HEALTH_EVERY_KEY "every"
 #define HEALTH_GREEN_KEY "green"
 #define HEALTH_RED_KEY "red"
@@ -57,44 +50,9 @@ extern unsigned int default_health_enabled;
 #define HEALTH_DELAY_KEY "delay"
 #define HEALTH_OPTIONS_KEY "options"
 
-/*
-typedef struct silencer {
-    char *alarms;
-    SIMPLE_PATTERN *alarms_pattern;
-
-    char *hosts;
-    SIMPLE_PATTERN *hosts_pattern;
-
-    char *contexts;
-    SIMPLE_PATTERN *contexts_pattern;
-
-    char *charts;
-    SIMPLE_PATTERN *charts_pattern;
-
-    char *families;
-    SIMPLE_PATTERN *families_pattern;
-
-    struct silencer *next;
-} SILENCER;
-*/
-
-typedef enum silence_type {
-    STYPE_NONE,
-    STYPE_DISABLE_ALARMS,
-    STYPE_SILENCE_NOTIFICATIONS
-} SILENCE_TYPE;
-
-typedef struct silencers {
-    int all_alarms;
-    SILENCE_TYPE stype;
-    SILENCER *silencers;
-} SILENCERS;
-
 #define HEALTH_SILENCERS_MAX_FILE_LEN 10000
-SILENCERS *silencers;
+
 char *silencers_filename;
-void health_silencers_add(SILENCER *silencer);
-SILENCER * health_silencers_addparam(SILENCER *silencer, char *key, char *value);
 
 extern void health_init(void);
 extern void *health_main(void *ptr);
@@ -142,9 +100,6 @@ extern void health_reload_host(RRDHOST *host);
 extern void health_alarm_log_free(RRDHOST *host);
 
 extern void health_alarm_log_free_one_nochecks_nounlink(ALARM_ENTRY *ae);
-
-extern int health_silencers_json_read_callback(JSON_ENTRY *e);
-extern SILENCER *create_silencer(void);
 
 extern void *health_cmdapi_thread(void *ptr);
